feat: 전적검색 API 구현
소환사명으로 요청 시 Slice된 전적정보와 챔피언 통계를 함께 리턴

+++ src/main/java/com/immersion/riot/match/infra/dto/ParticipantDto.java
@@ -34,7 +34,7 @@ public record ParticipantDto(
 
     public double caculateKDA() {
         if (deaths == 0) {
-            return (double) (kills + assists);
+            return (kills + assists);
         }
         return Math.round((kills + assists) / (double) deaths * 100) / 100.0;
     }
